Use authenticated username as prefix instead of hashed session key
With SESSION_COOKIE_HTTPONLY=True, the session key is marked httponly and therefore is not accessible to JavaScript / xhr. This is Django's sensible default.

We could use a hashed CSRF token. Even though Django 1.6+ has a CSRF_COOKIE_HTTPONLY option, it wouldn't really work with Fine Uploader, which probably needs to access the cookie in JavaScript. But this doesn't really need to be something secure, and would require an md5 hasher on the client. So we just use the username, which the client probably already has.

Optional middleware can provide this value in a cookie upon login.

Disables anonymous uploads, for now. Anonymous uploads would probably require something stored in the session.

In the long term, a better solution is to dispense with all this prefix stuff and use the S3 ETag to verify the user's upload hasn't been subsequently replaced. This would require Fine Uploader to provide that value to the success callback.